requirements.txt
python-dotenv==0.19.2
PyPDF2~=2.10.2